+++ src/components/Sections/Main/HeroSection.vue
