+++ src/components/Sections/Main/ProjectsSection.vue
