+++ src/pages/Home.vue
@@ -1,9 +1,9 @@
 <script setup lang="ts">
 import TheNavbar from './../components/_NavigationBar.vue';
-import HeroSection from './../components/HeroSection.vue';
-import AboutSection from './../components/AboutSection.vue';
-import SkillsSection from './../components/SkillsSection.vue';
-import ProjectsSection from './../components/ProjectsSection.vue';
+import HeroSection from '../components/Sections/Main/HeroSection.vue';
+import AboutSection from '../components/Sections/Main/AboutSection.vue';
+import SkillsSection from '../components/Sections/Main/SkillsSection.vue';
+import ProjectsSection from '../components/Sections/Main/ProjectsSection.vue';
 import ContactSection from './../components/ContactSection.vue';
 </script>
